requirements.txt
appdirs==1.4.3
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.3
click==6.7
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
futures==3.1.1
gunicorn==19.8.1
idna==2.5
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
packaging==16.8
PyMySQL==0.8.1
pyparsing==2.2.0
python-craigslist==1.0.3
python-dotenv==0.8.2
requests==2.17.3
requests-toolbelt==0.8.0
six==1.10.0
slackclient==1.0.5
SQLAlchemy==1.1.10
urllib3==1.21.1
websocket-client==0.40.0
Werkzeug==0.14.1
